WIP
Telling Stories with your commits
===

<ByLine/>

When was the last time you read a commit message? When was
the last time you looked at a individual commit?

> The reason for time, is so that things don't happen all
at once.

When we build software we don't build it all at once. We
build it in pieces, each part building on what came before.
These are more then just whispers, temporary, ephemeral,
that build into a mighty roar; these commits we make can
last forever. Preserved for the life of the project, inside
of our git histories.

Why not take some care and spiff them up a bit? Commit
messages are to often ignored. Serifs ontop of the actual
work that most would prefer Sans. So I made this, a way
to see into a website, how it got made, behind the curtain.

A series of commits that tell the story of how a website
got made.