[feature] generate commit log
Parsing my Git Log
==================

This is the heart of our application, and our first
non-boilerplate commit. We need content for our
website and that content is going to come mainly
from our git log.

Because we've SSRed everything, we're going to pull
the entire git log for our repo at build time.
Throw that into a json file that we can use to build
our page later on.

Most of the information could be parsed with `git`'s
`--pretty=format` flag to convert the git-log into
a standard json object. Extra care was needed to parse
the commit message, and patch diffs because those
needed to be sanatized before we could serialize
them. The diffs were further parsed with
[`parse-diff`][1]

[1]: github.com/sergeyt/parse-diff