[feature] GitHistory component
Render Git History on the Page
===

The main purpose of this app is to represent this repo's
git history on a page. These components do the work of
making that so.  A git history is made of of one or more
git commits, git commits have a summary (the first line of
a commit message), a optional body (any additional lines
of the commit message), and a set of git patches. Each
patch consists of chunks, that operate on specific files
and specify the changes that were made line by line.